Added player attack when looking at "enemy"
currently the distance is 2 away from the player but this can definitely be changed later. I also want to get it to be a cone in front of the player but I'm not sure how to do that yet. I also fixed the enemy seeking so it doesn't go towards the player's feet