Finished Player Dodge/Roll
The player now moves a small distance based on a timer when dodging. The player moves backwards when they press the dodge button if they are in the idle state/standing still. I also added in a NULL state to the Movement FSM for when the player is in the TAKEDAMAGE and DEAD states so that they cannot act when in those states in the Combat FSM